feat(local-only): implement support for tags
Usually, I commit based on the changes made locally, even partial implementations, but with this, this is kind of a mess, as commits were random asf, as most of the implementation has been done from my college lab (CRT classes 🤠☝️). And I was switching between stuff, so commit messages were a bit random. To cut all that crap, I had to squash everything; therefore, this commit.

This feature implementation isn't completed yet, but it's good enough for a preview. Remote implementation isn't completely implemented yet, so there's some work to do before a new release. Alongside this, there are a couple of other improvements that will be in the main release or new preview builds later.